Turbopack: organize tree shaking code
