+++ turbopack/crates/turbopack-ecmascript/src/tree_shake/part/mod.rs
@@ -1,0 +1,2 @@
+pub mod chunk_item;
+pub mod module;
