+++ turbopack/crates/turbopack-ecmascript/src/tree_shake/part/module.rs
@@ -11,10 +11,6 @@ use turbopack_core::{
     resolve::{ExportUsage, ModulePart},
 };
 
-use super::{
-    SplitResult, chunk_item::EcmascriptModulePartChunkItem, get_part_id, part_of_module,
-    split_module,
-};
 use crate::{
     AnalyzeEcmascriptModuleResult, EcmascriptAnalyzable, EcmascriptModuleAsset,
     EcmascriptModuleAssetType, EcmascriptModuleContent, EcmascriptModuleContentOptions,
@@ -25,7 +21,10 @@ use crate::{
         FollowExportsResult, analyse_ecmascript_module, esm::FoundExportType, follow_reexports,
     },
     side_effect_optimization::facade::module::EcmascriptModuleFacadeModule,
-    tree_shake::{Key, side_effect_module::SideEffectsModule},
+    tree_shake::{
+        Key, SplitResult, get_part_id, part::chunk_item::EcmascriptModulePartChunkItem,
+        part_of_module, side_effects::module::SideEffectsModule, split_module,
+    },
 };
 
 /// A reference to part of an ES module.
@@ -42,6 +41,7 @@ impl EcmascriptParsable for EcmascriptModulePartAsset {
     #[turbo_tasks::function]
     fn failsafe_parse(&self) -> Result<Vc<ParseResult>> {
         let split_data = split_module(*self.full_module);
+        assert_ne!(self.part, ModulePart::Facade);
         Ok(part_of_module(split_data, self.part.clone()))
     }
     #[turbo_tasks::function]
