+++ turbopack/crates/turbopack-ecmascript/src/tree_shake/side_effects/module.rs
@@ -15,11 +15,11 @@ use crate::{
     EcmascriptModuleAsset,
     chunk::{EcmascriptChunkPlaceable, EcmascriptExports},
     export::EcmascriptEvaluation,
-    tree_shake::chunk_item::SideEffectsModuleChunkItem,
+    tree_shake::side_effects::chunk_item::SideEffectsModuleChunkItem,
 };
 
 #[turbo_tasks::value]
-pub(super) struct SideEffectsModule {
+pub struct SideEffectsModule {
     /// Original module
     pub module: ResolvedVc<EcmascriptModuleAsset>,
     /// The part of the original module that is the binding
